// src/appointment/core/enums/condition-type-enum.ts
export enum CONDITION_TYPE {
    EQUALS = "==",
    GREATER_THAN = ">",
    LESS_THAN = "<",
    GREATER_THAN_OR_EQUAL_TO = ">=",
    LESS_THAN_OR_EQUAL_TO = "<=",
    DIFFERENT_THAN = "!=",
    LIKE = "like",
    IN = "in"
}


export enum CONDITION_VALUE {
    "==" = "Igual",
    ">" = "Mayor",
    "<" = "Menor",
    ">=" = "M ≥",
    "<=" = "M ≤",
    "!=" = "Diferente",
    "like" = "Buscar",
    "in" = "Estos"
}